Move Dna functionality to Dna class
